Added funny app icon, created UI and error handling.

+++ app/src/main/java/com/plcoding/weatherapp/domain/weather/WeatherData.kt
@@ -8,5 +8,5 @@ data class WeatherData(
     val pressure: Double,
     val windSpeed: Double,
     val weatherType: WeatherType,
-    val humidity: Int
+    val humidity: Double
 )
